Update classes with JPA Mapping and Dockerfile

+++ src/main/java/com/projetosoftwareteusz/teusz/repository/PedidoRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.projetosoftwareteusz.teusz.entities.Pedido;
 
 @Repository
-public interface PedidoRepository extends JpaRepository<Pedido, Long> {
+public interface PedidoRepository extends JpaRepository<Pedido, Integer> {
 
-}   
+}
